Rework import locations.
- Creates sewer/http_providers/ as a place to put http-01 AuthProviders.
- Creates auth.py which holds the new BaseAuthProvider class.
- Leaves BaseDns at the same import path for backwards compatibility.
- Exposes BaseHttp as sewer.BaseHttp for subclasses.
